Added login.cgi
It can now access the database via the DAO methods. For the moment you need to dump the DAO, Data and Utilities folders and their contents into C:\Apache24\htdocs for the cgi scripts to be able to call the DAO methods. Do not use objects, it won't work. Put all your object values into hashes.